moved about the data config
